+++ lib/src/partials/partials.barrel.dart
@@ -4,4 +4,5 @@ export 'partial_box_shadow.dart';
 export 'partial_outline_input_border.dart';
 export 'partial_underline_input_border.dart';
 export 'partial_no_input_border.dart';
+export 'partial_input_border_side.dart';
 export 'partial_shadow.dart';